requirements.txt
ordereddict
psycopg2

#Uncomment the following to use memcached
#python-memcached

#The following are for the gazetteer
-e git+https://github.com/cga-harvard/geopy.git#egg=geopy
geojson
#https://bitbucket.org/jespern/django-piston/get/5b3ef294fe89.zip
-e git+https://github.com/cga-harvard/django-piston.git#egg=django-piston
python-dateutil==1.5 #For Python < 3.0
datautil
jdcal


#Uncomment the following to use the django debug toolbar
#-e git+https://github.com/robhudson/django-debug-toolbar.git#egg=django-debug-toolbar

boto
-e git+https://github.com/deschler/django-modeltranslation#egg=django-modeltranslation

#Uncomment the following to use memcached
#python-memcached

#For scheduled tasks
django-celery
django-kombu

#For mapnotes
pytz
-e git+https://github.com/cga-harvard/vectorformats.git#egg=vectorformats

#For autocomplete
django-autocomplete-light==1.4.5
unidecode==0.04.14

#For haystack
django-haystack
pyelasticsearch

#ArcGIS REST services
-e git+https://github.com/Esri/arcpy-server-util-rest.git@hg/default#egg=arcrest
pyproj

#security
defusedxml